app/requirements.txt
Django>=5.0.6,<6
gunicorn>=22.0,<23
psycopg2>=2.9.9,<3
hiredis>=2.3.2,<3
redis>=5.0.4,<6
django-redis-sessions>=0.6.2,<1
djangorestframework>=3.15.2,<4
